# stockbuddy-backend-main/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StockBuddy API",
    "description": "Comprehensive API for StockBuddy - Inventory, Transaction, and Event Management System with multi-tenant architecture, JWT authentication, and extensive business logic for managing members, vendors, banks, transactions, inventory records, events, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "StockBuddy API Support",
      "email": "support@stockbuddy.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base Path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint. Include as: Authorization: Bearer <token>"
      }
    },
    "schemas": {
...
      "PersonalEvent": {
        "type": "object",
        "description": "User's personal calendar event with optional recurrence",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique event identifier"
          },
          "title": {
            "type": "string",
            "description": "Event title/name"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "description": "Event start date and time"
          },
          "repeat": {
            "type": "string",
            "enum": [
              "none",
              "daily",
              "weekly",
              "monthly",
              "yearly"
            ],
            "description": "Recurrence pattern"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get Event Categories",
        "description": "Get available event categories and offerings including sports and entertainment options.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Event categories retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "categories": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "sports": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/personal-events": {
      "get": {
        "tags": [
          "Personal Events"
        ],
        "summary": "List Personal Events",
        "description": "Get user's personal events with optional date range filtering.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start date filter (ISO format)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date filter (ISO format)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personal events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PersonalEvent"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Personal Events"
        ],
        "summary": "Create Personal Event",
        "description": "Create a new personal event with optional recurrence pattern.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "start_time"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Team Meeting"
                  },
                  "description": {
                    "type": "string",
                    "example": "Weekly team sync"
                  },
                  "start_time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-15T10:00:00Z"
                  },
                  "end_time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-15T11:00:00Z"
                  },
                  "repeat": {
                    "type": "string",
                    "enum": [
                      "none",
                      "daily",
                      "weekly",
                      "monthly",
                      "yearly"
                    ],
                    "default": "none"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Personal event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PersonalEvent"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}